requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncer==0.0.17
attrs==26.1.0
bidict==0.23.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chainlit==2.10.1
charset-normalizer==3.4.6
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.135.3
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-agno==0.57.0
opentelemetry-instrumentation-alephalpha==0.57.0
opentelemetry-instrumentation-anthropic==0.57.0
opentelemetry-instrumentation-bedrock==0.57.0
opentelemetry-instrumentation-chromadb==0.57.0
opentelemetry-instrumentation-cohere==0.57.0
opentelemetry-instrumentation-crewai==0.57.0
opentelemetry-instrumentation-google-generativeai==0.57.0
opentelemetry-instrumentation-groq==0.57.0
opentelemetry-instrumentation-haystack==0.57.0
opentelemetry-instrumentation-lancedb==0.57.0
opentelemetry-instrumentation-langchain==0.57.0
opentelemetry-instrumentation-llamaindex==0.57.0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-marqo==0.57.0
opentelemetry-instrumentation-mcp==0.57.0
opentelemetry-instrumentation-milvus==0.57.0
opentelemetry-instrumentation-mistralai==0.57.0
opentelemetry-instrumentation-ollama==0.57.0
opentelemetry-instrumentation-openai==0.57.0
opentelemetry-instrumentation-openai-agents==0.57.0
opentelemetry-instrumentation-pinecone==0.57.0
opentelemetry-instrumentation-qdrant==0.57.0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-replicate==0.57.0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sagemaker==0.57.0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-together==0.57.0
opentelemetry-instrumentation-transformers==0.57.0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-vertexai==0.57.0
opentelemetry-instrumentation-voyageai==0.57.0
opentelemetry-instrumentation-watsonx==0.57.0
opentelemetry-instrumentation-weaviate==0.57.0
opentelemetry-instrumentation-writer==0.57.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.5.1
opentelemetry-util-http==0.61b0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2026.1.post1
pywin32==311
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.4
starlette==1.0.0
syncer==2.0.3
tenacity==9.1.4
toml==0.10.2
tomli==2.4.1
tornado==6.5.5
tqdm==4.67.3
traceloop-sdk==0.57.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchdog==6.0.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.0